pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sksecurity"
version = "1.2.1"
description = "Enterprise-grade security for AI agent ecosystems"
readme = "README.md"
license = "GPL-3.0-or-later"
requires-python = ">=3.9"
authors = [
    {name = "smilinTux Team", email = "team@smilintux.org"},
]
keywords = [
    "security", "ai", "agents", "scanning", "threat-intelligence",
    "monitoring", "quarantine", "cybersecurity", "soc",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Information Technology",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: System :: Monitoring",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
    "Environment :: Console",
    "Environment :: Web Environment",
]
dependencies = [
    "click>=8.0.0",
    "pyyaml>=6.0",
    "flask>=2.3.0",
    "flask-cors>=4.0.0",
    "sqlalchemy>=2.0.0",
    "cryptography>=41.0.0",
    "requests>=2.31.0",
    "psutil>=5.9.0",
    "watchdog>=3.0.0",
    "email-validator>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[project.urls]
Homepage = "https://sksecurity.io"
Repository = "https://github.com/smilinTux/sksecurity"
"Bug Tracker" = "https://github.com/smilinTux/sksecurity/issues"

[project.scripts]
sksecurity = "sksecurity.cli:main"

[tool.setuptools.packages.find]
include = ["sksecurity*"]
exclude = ["tests*"]

[tool.setuptools.package-data]
sksecurity = ["data/*.json", "data/*.yml", "templates/*.html", "static/*"]
